+++ src/components/index.ts
@@ -44,7 +44,7 @@ export {
   default as CollapsibleListEntry,
   type CollapsibleListEntryContentProps
 } from './CollapsibleList/CollapsibleListEntry/CollapsibleListEntry'
-export { default as LinearProgressWithLabel } from './LinearProgressWithLabel/LinearProgressWithLabel'
-export { default as BorderLinearProgress } from './BorderLinearProgress/BorderLinearProgress'
+export { default as LinearProgressWithLabel } from './StyledLinearProgress/LinearProgressWithLabel'
+export { StyledLinearProgress } from './StyledLinearProgress/StyledLinearProgress'
 
 export * from './Nodes' // This stays as is because it's not following the Component/Component structure
